Rewrite build and deploy in typescript
